Conflicts:
	plugins/org.eclipse.etrice.generator.c/xtend-gen/org/eclipse/etrice/generator/c/gen/CExtensions.java
	plugins/org.eclipse.etrice.generator.java/xtend-gen/org/eclipse/etrice/generator/java/gen/ActorClassGen.java
	plugins/org.eclipse.etrice.generator.java/xtend-gen/org/eclipse/etrice/generator/java/gen/JavaExtensions.java
	plugins/org.eclipse.etrice.generator.java/xtend-gen/org/eclipse/etrice/generator/java/gen/SubSystemClassGen.java
	plugins/org.eclipse.etrice.generator/xtend-gen/org/eclipse/etrice/generator/generic/ConfigExtension.java
	plugins/org.eclipse.etrice.generator/xtend-gen/org/eclipse/etrice/generator/generic/ProcedureHelpers.java
	plugins/org.eclipse.etrice.generator/xtend-gen/org/eclipse/etrice/generator/generic/RoomExtensions.java
	plugins/org.eclipse.etrice.generator/xtend-gen/org/eclipse/etrice/generator/generic/TypeHelpers.java
	plugins/org.eclipse.etrice.ui.runtime/contents/org.eclipse.etrice.modellib.c.zip
	plugins/org.eclipse.etrice.ui.runtime/contents/org.eclipse.etrice.modellib.java.zip
	plugins/org.eclipse.etrice.ui.runtime/contents/org.eclipse.etrice.runtime.c.zip
	plugins/org.eclipse.etrice.ui.runtime/contents/org.eclipse.etrice.runtime.java.zip
	plugins/org.eclipse.etrice.ui.runtime/contents/org.eclipse.etrice.tutorials.zip
	runtime/org.eclipse.etrice.modellib.java/src-gen/org/eclipse/etrice/modellib/Socket/SocketClient.java
	runtime/org.eclipse.etrice.modellib.java/src-gen/room/basic/service/logging/ALogService.java
	runtime/org.eclipse.etrice.modellib.java/src-gen/room/basic/service/tcp/ATcpClient.java
	runtime/org.eclipse.etrice.modellib.java/src-gen/room/basic/service/tcp/ATcpServer.java

+++ plugins/org.eclipse.etrice.core.config/src-gen/org/eclipse/etrice/core/config/impl/IntLiteralImpl.java
@@ -34,7 +34,7 @@ public class IntLiteralImpl extends NumberLiteralImpl implements IntLiteral
    * @generated
    * @ordered
    */
-  protected static final int VALUE_EDEFAULT = 0;
+  protected static final long VALUE_EDEFAULT = 0L;
 
   /**
    * The cached value of the '{@link #getValue() <em>Value</em>}' attribute.
@@ -44,7 +44,7 @@ public class IntLiteralImpl extends NumberLiteralImpl implements IntLiteral
    * @generated
    * @ordered
    */
-  protected int value = VALUE_EDEFAULT;
+  protected long value = VALUE_EDEFAULT;
 
   /**
    * <!-- begin-user-doc -->
@@ -72,7 +72,7 @@ public class IntLiteralImpl extends NumberLiteralImpl implements IntLiteral
    * <!-- end-user-doc -->
    * @generated
    */
-  public int getValue()
+  public long getValue()
   {
     return value;
   }
@@ -82,9 +82,9 @@ public class IntLiteralImpl extends NumberLiteralImpl implements IntLiteral
    * <!-- end-user-doc -->
    * @generated
    */
-  public void setValue(int newValue)
+  public void setValue(long newValue)
   {
-    int oldValue = value;
+    long oldValue = value;
     value = newValue;
     if (eNotificationRequired())
       eNotify(new ENotificationImpl(this, Notification.SET, ConfigPackage.INT_LITERAL__VALUE, oldValue, value));
@@ -117,7 +117,7 @@ public class IntLiteralImpl extends NumberLiteralImpl implements IntLiteral
     switch (featureID)
     {
       case ConfigPackage.INT_LITERAL__VALUE:
-        setValue((Integer)newValue);
+        setValue((Long)newValue);
         return;
     }
     super.eSet(featureID, newValue);
